+++ examples/onnxrt/image_recognition/vgg16/quantization/ptq_static/main.py
@@ -139,7 +139,7 @@ class Dataloader:
 
     def _preprpcess(self, src):
         with Image.open(src) as image:
-            image = np.array(image.convert('RGB')).astype(np.float32)
+            image = np.array(image.convert('RGB'))
             
             height, width = image.shape[0], image.shape[1]
             scale = self.resize_side / width if height > width else self.resize_side / height
